#3941 add integration tests for transpile
