+++ .codacy.yml
@@ -6,4 +6,4 @@
 #  package name contains capital letter and such names are conventional.
 ---
 exclude_paths:
-  - "eo-maven-plugin/src/test/java/org/eolang/maven/RegisterMojoIT.java"
+  - "eo-maven-plugin/src/test/java/org/eolang/maven/TranspileMojoIT.java"
