+++ eo-maven-plugin/src/test/java/org/eolang/maven/TranspileMojoTest.java
@@ -6,9 +6,7 @@ package org.eolang.maven;
 
 import com.yegor256.Mktmp;
 import com.yegor256.MktmpResolver;
-import com.yegor256.farea.Farea;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -98,75 +96,6 @@ final class TranspileMojoTest {
                     Matchers.hasKey("target/generated/EOcom/EOexample/EOfoo.java")
                 )
             )
-        );
-    }
-
-    @Test
-    void transpilesWithPackage(@Mktmp final Path temp)
-        throws Exception {
-        new Farea(temp).together(
-            f -> {
-                f.clean();
-                f.files().file("src/main/eo/one/foo.eo").write(
-                    String.join(
-                        "\n",
-                        "+package one",
-                        "+unlint object-has-data",
-                        "",
-                        "# no comments.",
-                        "[] > foo",
-                        "  QQ.io.stdout > @",
-                        "    \"Hello, world!\\n\"",
-                        ""
-                    ).getBytes(StandardCharsets.UTF_8)
-                );
-                f.build()
-                    .plugins()
-                    .appendItself()
-                    .execution()
-                    .goals("register", "parse", "shake", "transpile");
-                f.exec("process-sources");
-            }
-        );
-        MatcherAssert.assertThat(
-            "the .java file is generated",
-            temp.resolve("target/generated-sources/EOone/EOfoo.java").toFile().exists(),
-            Matchers.is(true)
-        );
-        MatcherAssert.assertThat(
-            "the package-info.java file contains the right package name",
-            Files.readString(
-                temp.resolve("target/generated-sources/EOone/package-info.java"),
-                StandardCharsets.UTF_8
-            ),
-            Matchers.containsString("package EOone;")
-        );
-    }
-
-    @Test
-    void transpilesSimpleApp(@Mktmp final Path temp, @RandomProgram final String prog)
-        throws Exception {
-        new Farea(temp).together(
-            f -> {
-                f.clean();
-                f.files().file("src/main/eo/foo.eo").write(prog.getBytes());
-                f.build()
-                    .plugins()
-                    .appendItself()
-                    .execution()
-                    .goals("register", "parse", "shake", "transpile");
-                f.exec("process-sources");
-            }
-        );
-        MatcherAssert.assertThat(
-            "the .java file is re-generated",
-            temp.resolve("target/generated-sources/EOfoo.java").toFile().exists(),
-            Matchers.is(true)
-        );
-        MatcherAssert.assertThat(
-            "the package-info.java file contains the right package name",
-            temp.resolve("target/generated-sources/package-info.java").toFile().exists(),
-            Matchers.is(false)
         );
     }
